s0 in extraneighbor is now scaled by A. implemened bisection for discovering roots of mech eq. cond.

+++ apps/extraneighbor/extraneighbor.cpp
@@ -36,7 +36,7 @@ void ExtraNeighbor::notifyObserver(const Subjects &subject)
 
         SurfaceReaction &r = solver().surfaceReaction(x, y);
 
-        if (csc.type == ChangeTypes::Double)
+        if (csc.type == ChangeTypes::Single)
         {
             const int &value = csc.value;
